adding calc_params_per_spot for zarr+restructure folders
