+++ xdbit_funcs/__init__.py
@@ -6,6 +6,3 @@ from . import images as im
 from . import utils as utils
 from . import datasets as data
 from . import preprocessing as pp
-
-
-
