+++ xdbit_funcs/datasets/__init__.py
@@ -1,1 +1,2 @@
 from .get import *
+from .objects import *
