+++ xdbit_funcs/plotting/__init__.py
@@ -5,6 +5,7 @@ from .qc import *
 from .go import *
 from .interactive import *
 from .spatial_class import MultiSpatialPlot as spatial
+from .spatial_class import ImageData
 from .color_palettes import *
